refactor(api): move competence-repository
