+++ api/db/seeds/data/common/tooling/learning-content.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import * as competenceRepository from '../../../../../lib/infrastructure/repositories/competence-repository.js';
+import * as competenceRepository from '../../../../../src/shared/infrastructure/repositories/competence-repository.js';
 import * as challengeRepository from '../../../../../lib/infrastructure/repositories/challenge-repository.js';
 import { skillDatasource } from '../../../../../lib/infrastructure/datasources/learning-content/skill-datasource.js';
