+++ api/lib/domain/services/placement-profile-service.js
@@ -7,7 +7,7 @@ import * as assessmentRepository from '../../../src/shared/infrastructure/reposi
 import * as skillRepository from '../../infrastructure/repositories/skill-repository.js';
 import * as assessmentResultRepository from '../../infrastructure/repositories/assessment-result-repository.js';
 import * as knowledgeElementRepository from '../../infrastructure/repositories/knowledge-element-repository.js';
-import * as competenceRepository from '../../infrastructure/repositories/competence-repository.js';
+import * as competenceRepository from '../../../src/shared/infrastructure/repositories/competence-repository.js';
 import * as scoringService from './scoring/scoring-service.js';
 import { CertificationVersion } from '../../../src/shared/domain/models/CertificationVersion.js';
