+++ api/lib/infrastructure/repositories/campaign-participation-result-repository.js
@@ -1,7 +1,7 @@
 import { CampaignParticipationResult } from '../../domain/models/CampaignParticipationResult.js';
 import * as campaignParticipationRepository from './campaign-participation-repository.js';
 import * as campaignRepository from './campaign-repository.js';
-import * as competenceRepository from './competence-repository.js';
+import * as competenceRepository from '../../../src/shared/infrastructure/repositories/competence-repository.js';
 import * as areaRepository from './area-repository.js';
 import * as assessmentRepository from '../../../src/shared/infrastructure/repositories/assessment-repository.js';
 import * as knowledgeElementRepository from './knowledge-element-repository.js';
