+++ api/lib/infrastructure/repositories/learning-content-repository.js
@@ -4,7 +4,7 @@ import { NoSkillsInCampaignError, NotFoundError } from '../../domain/errors.js';
 import * as tubeRepository from './tube-repository.js';
 import * as thematicRepository from './thematic-repository.js';
 import * as campaignRepository from './campaign-repository.js';
-import * as competenceRepository from './competence-repository.js';
+import * as competenceRepository from '../../../src/shared/infrastructure/repositories/competence-repository.js';
 import * as areaRepository from './area-repository.js';
 import * as frameworkRepository from './framework-repository.js';
 import * as skillRepository from './skill-repository.js';
