+++ api/lib/infrastructure/repositories/participant-result-repository.js
@@ -2,7 +2,7 @@ import { knex } from '../../../db/knex-database-connection.js';
 import _ from 'lodash';
 import { Assessment } from '../../domain/models/index.js';
 import { AssessmentResult } from '../../domain/read-models/participant-results/AssessmentResult.js';
-import * as competenceRepository from './competence-repository.js';
+import * as competenceRepository from '../../../src/shared/infrastructure/repositories/competence-repository.js';
 import * as answerRepository from '../../../src/evaluation/infrastructure/repositories/answer-repository.js';
 import * as challengeRepository from './challenge-repository.js';
 import * as areaRepository from './area-repository.js';
