+++ api/lib/infrastructure/repositories/target-profile-for-admin-repository.js
@@ -6,7 +6,7 @@ import { LOCALE } from '../../domain/constants.js';
 const { FRENCH_FRANCE } = LOCALE;
 
 import * as areaRepository from './area-repository.js';
-import * as competenceRepository from './competence-repository.js';
+import * as competenceRepository from '../../../src/shared/infrastructure/repositories/competence-repository.js';
 import * as thematicRepository from './thematic-repository.js';
 import * as tubeRepository from './tube-repository.js';
 import * as skillRepository from './skill-repository.js';
