+++ api/scripts/certification/generate-certification-test-statistics.js
@@ -9,7 +9,7 @@ import originalFp from 'lodash/fp.js';
 const fp = originalFp.convert({ cap: false });
 import bluebird from 'bluebird';
 import { knex, disconnect } from '../../db/knex-database-connection.js';
-import * as competenceRepository from '../../lib/infrastructure/repositories/competence-repository.js';
+import * as competenceRepository from '../../src/shared/infrastructure/repositories/competence-repository.js';
 import * as placementProfileService from '../../lib/domain/services/placement-profile-service.js';
 import * as certificationChallengeService from '../../lib/domain/services/certification-challenges-service.js';
 import { LOCALE } from '../../lib/domain/constants.js';
