+++ api/scripts/tooling/tooling.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import bluebird from 'bluebird';
 import * as skillRepository from '../../lib/infrastructure/repositories/skill-repository.js';
-import * as competenceRepository from '../../lib/infrastructure/repositories/competence-repository.js';
+import * as competenceRepository from '../../src/shared/infrastructure/repositories/competence-repository.js';
 import * as challengeRepository from '../../lib/infrastructure/repositories/challenge-repository.js';
 import * as campaignRepository from '../../lib/infrastructure/repositories/campaign-repository.js';
 import { logger } from '../../lib/infrastructure/logger.js';
