+++ api/src/evaluation/infrastructure/repositories/competence-evaluation-repository.js
@@ -1,9 +1,9 @@
 import _ from 'lodash';
-import { NotFoundError } from '../../../../lib/domain/errors.js';
 import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
 import { knex } from '../../../../db/knex-database-connection.js';
 import { CompetenceEvaluation } from '../../domain/models/CompetenceEvaluation.js';
 import { Assessment } from '../../../shared/domain/models/Assessment.js';
+import { NotFoundError } from '../../../shared/domain/errors.js';
 
 const save = async function ({ competenceEvaluation, domainTransaction = DomainTransaction.emptyTransaction() }) {
   const knexConn = domainTransaction.knexTransaction || knex;
